+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - feat: use logging instead of simple print
 - feat: introduce a global state and send messages depending on the state, improve logging
+- ci: add python linting workflow
 
 ### Bug Fixes