+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - feat: introduce a global state and send messages depending on the state, improve logging
 - ci: add python linting workflow
 - stlye: fix flake8 findings
+- stlye: fix pylint findings
 
 ### Bug Fixes