+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - feat: use logging instead of simple print
 - feat: introduce a global state and send messages depending on the state, improve logging
 - ci: add python linting workflow
+- ci: set max line length for flake8 python linting workflow
 - stlye: fix flake8 findings
 - stlye: fix pylint findings